+++ R/worldplot.R
@@ -29,7 +29,8 @@
 #' @importFrom dplyr "%>%" left_join select filter mutate relocate
 #' @importFrom ggplot2 ggplot geom_sf theme labs scale_fill_viridis_c coord_sf xlab ylab ggtitle
 #'                     aes unit element_text element_blank element_rect geom_text ggsave
-#' @importFrom sf st_centroid st_coordinates
+#' @importFrom sf st_centroid st_coordinates st_union
+#' @importFrom ggfx with_shadow
 #'
 #' @examples
 #' data(testdata1b)
@@ -56,6 +57,12 @@ worldplot <- function(data,
     mutate(iso_a2 = ifelse(name %in% c("Indian Ocean Ter." , "Ashmore and Cartier Is."), -99, iso_a2_eh),
            iso_a3 = ifelse(name %in% c("Indian Ocean Ter." , "Ashmore and Cartier Is."), -99, iso_a3_eh)) %>%
     select(name, iso_a2, iso_a3, geometry)
+
+  #Cyprus adjustment
+  cyp <- subset(map_df0, name %in% c("Cyprus", "N. Cyprus"))
+  cyp2 <- st_union(cyp[1, "geometry"], cyp[2,"geometry"])
+  map_df0[map_df0$iso_a2 == "CY", "geometry"] <- cyp2
+  # end of cyprus adjustment
 
   simdata <- c()
 
@@ -97,7 +104,8 @@ worldplot <- function(data,
                                     countries.list = simdata$iso_a2[!is.na(simdata$MapFiller)])
 
     wplot <- wplot +
-      geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= 2/div, color= 'black', fontface= 'bold')
+      with_shadow(geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= 2/div, color= 'white', fontface= 'bold'),
+                  x_offset = 2, y_offset = 2, sigma = 1)
   }
 
   print(wplot)