+++ R/worldplotCat.R
@@ -5,6 +5,7 @@
 #' @inheritParams worldplot
 #' @param Categories categories labels to be plotted in the legend.
 #' @param na.as.category Treat \code{NA} as a separate category? If `\code{TRUE}, NA will also appear in the legend as one of the categories.
+#' @param palette_option Character string indicating the palette to be used. Available options range between "A" and "H". You can also enter a string with a colour for each category
 #'
 #' @return a map
 #' @export
@@ -13,8 +14,9 @@
 #' @importFrom countrycode countrycode
 #' @importFrom dplyr "%>%" left_join select select filter mutate relocate
 #' @importFrom ggplot2 ggplot geom_sf theme labs scale_fill_viridis_d coord_sf xlab ylab ggtitle
-#'                     aes unit element_text element_blank element_rect geom_text ggsave
-#' @importFrom sf st_centroid st_coordinates
+#'                     aes unit element_text element_blank element_rect geom_text ggsave scale_fill_manual
+#' @importFrom sf st_centroid st_coordinates st_union
+#' @importFrom ggfx with_shadow
 #'
 #' @examples
 #' data(testdata1b)
@@ -42,6 +44,12 @@ worldplotCat <- function(data,
     mutate(iso_a2 = ifelse(name %in% c("Indian Ocean Ter." , "Ashmore and Cartier Is."), -99, iso_a2_eh),
            iso_a3 = ifelse(name %in% c("Indian Ocean Ter." , "Ashmore and Cartier Is."), -99, iso_a3_eh)) %>%
     select(name, iso_a2, iso_a3, geometry)
+
+  #Cyprus adjustment
+  cyp <- subset(map_df0, name %in% c("Cyprus", "N. Cyprus"))
+  cyp2 <- st_union(cyp[1, "geometry"], cyp[2,"geometry"])
+  map_df0[map_df0$iso_a2 == "CY", "geometry"] <- cyp2
+  # end of cyprus adjustment
 
   simdata <- c()
 
@@ -72,11 +80,19 @@ worldplotCat <- function(data,
           panel.grid = element_blank(),
           panel.background = element_rect(fill = 'grey95'))+
     labs(fill= legendTitle)+
-    scale_fill_viridis_d(option = palette_option, begin= 0.3, na.value = 'grey80', direction= 1,
-                         labels= c(Categories, "NA"), na.translate = na.as.category)+
     coord_sf(xlim= longitude, ylim= latitude, expand= FALSE, label_axes = 'SW') +
     xlab('')+ ylab('')+
     ggtitle(title)
+
+  if (length(palette_option) == 1) {
+    wplot <- wplot +
+      scale_fill_viridis_d(option = palette_option, begin= 0.3, na.value = 'grey80', direction= 1,
+                         labels= c(Categories, "NA"), na.translate = na.as.category)
+  } else {
+    wplot <- wplot +
+      scale_fill_manual(values = palette_option, na.value="grey90", drop= F,
+                        labels = Categories)
+  }
 
   if (annote == TRUE) {
 
@@ -84,7 +100,8 @@ worldplotCat <- function(data,
                                     countries.list = simdata$iso_a2[!is.na(simdata$MapFiller)])
 
     wplot <- wplot +
-      geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= 2/div, color= 'black', fontface= 'bold')
+      with_shadow(geom_text(data= world_points, aes(x=X, y=Y,label= iso_a2), size= 2/div, color= 'white', fontface= 'bold'),
+                  x_offset = 2, y_offset = 2, sigma = 1)
   }
 
   print(wplot)