+++ tests/testthat/testCat2.R
@@ -7,3 +7,27 @@ worldplotCat(data = testdata1b,
              Categories = c("Low", "Average", "High"),
              legendTitle = "CAT",
              annote = TRUE)
+
+
+worldplotCat(data = testdata1b,
+             div = 1,
+             ColName = "VCat",
+             CountryName = "Cshort",
+             CountryNameType = "isoa2",
+             title = "Testing the function for Categorical variables",
+             Categories = c("Low", "Average", "High"),
+             legendTitle = "CAT",
+             annote = TRUE)
+
+worldplotCat(data = testdata1b,
+             div = 1,
+             ColName = "VCat",
+             CountryName = "Cshort",
+             CountryNameType = "isoa2",
+             title = "Testing the function, custom colours",
+             palette_option = c("#C3E2EA", "#58C0D0", "#256C91"),
+             Categories = c("Low", "Average", "High"),
+             legendTitle = "CAT",
+             annote = TRUE)
+
+