+++ NEWS.md
@@ -1,3 +1,10 @@
+# WorldMapR 0.1.1
+
+* Label's position has been improved for some countries in geometries_data
+* For categorical variables, the color palette can now be customized manually 
+* Improved labels' aesthetics
+
+
 # WorldMapR 0.1.0
 
 * Initial CRAN submission.